requirements.txt
azure-functions
azure-storage-blob
azure-servicebus
azure-eventhub
